+++ Proxii.Test/Integration/Interceptors/OnReturnInterceptorTest.cs
@@ -1,0 +1,19 @@
+﻿using Microsoft.VisualStudio.TestTools.UnitTesting;
+using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace Proxii.Test.Integration.Interceptors
+{
+    [TestClass]
+    public class OnReturnInterceptorTest
+    {
+        [TestMethod]
+        public void Integration_OnReturnInterceptor()
+        {
+            Assert.Fail();
+        }
+    }
+}