+++ Proxii.Test/e2e/ProxiiOnReturnTest.cs
@@ -1,0 +1,19 @@
+﻿using Microsoft.VisualStudio.TestTools.UnitTesting;
+using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace Proxii.Test.e2e
+{
+    [TestClass]
+    public class ProxiiOnReturnTest
+    {
+        [TestMethod]
+        public void Proxii_OnReturn()
+        {
+            Assert.Fail();
+        }
+    }
+}